:hammer: Updated port number handling for gRPC server
https://gitlab.sirehna.com/sirehna/xdyn/-/merge_requests/160#note_113573

Port numbers 0 to 1024 are reserved for privileged services and designated as well-known ports. This list of port numbers are specified in RFC 1700. The Dynamic and/or Private Ports are those from 49152 through 65535 (so they're probably safest). In any case, the port number should be within [0, 65535] (included).

+++ code/grpc/inc/AiryGRPC.hpp
@@ -5,6 +5,6 @@
 #define GRPC_INC_AIRYGRPC_HPP_
 
 class EnvironmentAndFrames;
-void run_xdyn_airy_server(const EnvironmentAndFrames& env, const int port_number);
+void run_xdyn_airy_server(const EnvironmentAndFrames& env, const short unsigned int port_number);
 
 #endif /* GRPC_INC_AIRYGRPC_HPP_ */
